+++ Source/DiabloUI/selok.cpp
@@ -30,7 +30,7 @@ void selok_Free()
 	vecSelOkDialog.clear();
 }
 
-void selok_Select(int /*value*/)
+void selok_Select(size_t /*value*/)
 {
 	selok_endMenu = true;
 }
